CRUD Django task (adding more functionalities to the blog)
From your knowledge of Python and Django. Extend the blog we built with the following functionalities: A register page (a new user can register, their details stored in database file). login (checks in the file and logs them in if they are already registered),reset password, logout, A comment section, you must be logged in to comment.